Specify the Sonarqube edition in Compute Engine to overcome restrictions
As part of SONAR-12081 a check has been added into the Compute Engine to prevent Community Edition versions of SonarQube from being able to submit branches. To work around this, the plugin now exposes a provider to the Compute Engine to make the engine believe that Developer edition is being run. To prevent the update center and footer panels showing the incorrect edition to users, the new provider has not been exposed to the SonarQube server component, thereby leaving the default Community Edition fall-back to execute.