#47: Fix NPE when branch is orphaned
Where a Pull Request no longer has a target branch - such as happens when the branch is deleted from SonarQube after the Pull Request is analysed - the `base` field in the branch information will be `null`. To prevent a `NullPointerException` being thrown in this scenario, an explicit check is being added for this field being `null` before attempting to extract the text value from it.